jai ajoute une colonne blocquee dans table users et jai mettre dans methode registre complete dans registrerUsercontroller que blocquee egal non blocquee et jai mettre dans contriller signal dans methodeupdate signal que si le signal >= 3 colonne blocquee egal blocquee

+++ app/Http/Controllers/Entry/LoginUserController.php
@@ -17,6 +17,9 @@ class LoginUserController extends Controller
         ];
         if(auth()->attempt($donnee)){
             $user=Auth::user();
+            if($user->blocquee=="blocquee"){
+                 return response()->json(['error' => 'Votre compte est bloqué'], 403);
+            }
             $user->tokens()->delete();
             $success['name']=$user->name;
             $success['token']=$user->createToken(request()->userAgent())->plainTextToken;
